# docs/referencia/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Applications Tracker API",
    "description": "\nAPI REST de **Applications Tracker**: seguimiento de solicitudes a puestos de trabajo.\n\n## Autenticación\n\nLa autenticación la gestiona SuperTokens (email y contraseña). Hay dos modos, y la\nAPI acepta los dos en todas las rutas protegidas:\n\n| Cliente | Modo | Cómo |\n|---|---|---|\n| Navegador (la SPA) | Cookies httpOnly | Login con la cabecera `st-auth-mode: cookie`. El navegador envía las cookies solo. |\n| Integraciones, scripts, Swagger | Bearer | Login con `st-auth-mode: header` (o sin esa cabecera). Se usa `Authorization: Bearer <st-access-token>`. |\n\n### Probar desde esta página\n\n1. Despliega **POST /auth/signin**, pulsa *Try it out* y ejecuta con un usuario existente.\n2. En *Response headers*, copia el valor de `st-access-token`.\n3. Pulsa **Authorize**, pega el token en *BearerAuth* y confirma.\n4. Las rutas con candado ya se pueden ejecutar. El token caduca a los **5 minutos**: renuévalo con `POST /auth/session/refresh` o repite el login.\n\n## Convenciones\n\n- **Errores de validación:** 422, con el detalle de cada campo en `detail`.\n- **Sin sesión:** 401 `{\"message\": \"unauthorised\"}`.\n- **Recursos de otro usuario:** 404, igual que si no existieran.\n- **Listados paginados:** `page` (desde 1) y `limit` (máximo 100) →\n  `{items, total, page, limit, pages}`.\n- **Fechas:** ISO 8601 en UTC.\n",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/reminders": {
      "get": {
        "tags": [
          "Reminders"
        ],
        "summary": "Listar recordatorios",
        "description": "Recordatorios del usuario (RF-52), por defecto solo los pendientes. Filtra\npor `application_id` para los de una solicitud concreta (su detalle, RF-52).",
        "operationId": "list_reminders_api_v1_reminders_get",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Página, empezando en 1.",
              "default": 1,
              "title": "Page"
            },
            "description": "Página, empezando en 1."
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "description": "Elementos por página.",
              "default": 20,
              "title": "Limit"
            },
            "description": "Elementos por página."
          },
          {
            "name": "application_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Solo los recordatorios de esta solicitud.",
              "title": "Application Id"
            },
            "description": "Solo los recordatorios de esta solicitud."
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "pending",
                "done",
                "dismissed",
                "all"
              ],
              "type": "string",
              "description": "`pending` (por defecto), `done`, `dismissed` o `all`.",
              "default": "pending",
              "title": "Status"
            },
            "description": "`pending` (por defecto), `done`, `dismissed` o `all`."
          },
          {
            "name": "due_before",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Due Before"
            }
          },
          {
            "name": "due_after",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Due After"
            }
          },
          {
            "name": "sort_by",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "due_at",
                "created_at"
              ],
              "type": "string",
              "default": "due_at",
              "title": "Sort By"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "asc",
                "desc"
              ],
              "type": "string",
              "default": "asc",
              "title": "Order"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ReminderRead_"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión, o el access token ha caducado o no es válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Reminders"
        ],
        "summary": "Crear un recordatorio",
        "description": "Crea un recordatorio, opcionalmente ligado a una solicitud (RF-50). 404 si\n`application_id` no existe o es de otro usuario. 409\n`reminders_limit_reached` al superar 500 recordatorios **pendientes**.",
        "operationId": "create_reminder_api_v1_reminders_post",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CookieAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReminderCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderRead"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión, o el access token ha caducado o no es válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "404": {
            "description": "No existe o pertenece a otro usuario (no se distingue).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Una regla de negocio impide la operación; ver `code`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reminders/{reminder_id}/complete": {
      "post": {
        "tags": [
          "Reminders"
        ],
        "summary": "Marcar un recordatorio como hecho",
        "description": "RF-51. 409 `reminder_not_pending` si ya estaba hecho o descartado.",
        "operationId": "complete_reminder_api_v1_reminders__reminder_id__complete_post",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "reminder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Reminder Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderRead"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión, o el access token ha caducado o no es válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "404": {
            "description": "No existe o pertenece a otro usuario (no se distingue).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Una regla de negocio impide la operación; ver `code`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reminders/{reminder_id}/dismiss": {
      "post": {
        "tags": [
          "Reminders"
        ],
        "summary": "Descartar un recordatorio",
        "description": "RF-51. 409 `reminder_not_pending` si ya estaba hecho o descartado.",
        "operationId": "dismiss_reminder_api_v1_reminders__reminder_id__dismiss_post",
        "security": [
          {
            "BearerAuth": []
          },
          {
            "CookieAuth": []
          }
        ],
        "parameters": [
          {
            "name": "reminder_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Reminder Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReminderRead"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión, o el access token ha caducado o no es válido.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UnauthorizedError"
                }
              }
            }
          },
          "404": {
            "description": "No existe o pertenece a otro usuario (no se distingue).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Una regla de negocio impide la operación; ver `code`.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail",
            "examples": [
              "Company not found"
            ]
          },
          "code": {
            "type": "string",
            "title": "Code",
            "examples": [
              "not_found"
            ]
          }
        },
        "type": "object",
        "required": [
          "detail",
          "code"
        ],
        "title": "ErrorResponse",
        "description": "Error de negocio. `code` es estable (contrato); `detail` es informativo."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Page_ReminderRead_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReminderRead"
            },
            "type": "array",
            "title": "Items",
            "description": "Elementos de la página pedida."
          },
          "total": {
            "type": "integer",
            "title": "Total",
            "description": "Total de elementos que cumplen el filtro."
          },
          "page": {
            "type": "integer",
            "title": "Page",
            "description": "Página devuelta, empezando en 1."
          },
          "limit": {
            "type": "integer",
            "title": "Limit",
            "description": "Tamaño de página pedido."
          },
          "pages": {
            "type": "integer",
            "title": "Pages",
            "description": "Número total de páginas (0 si no hay elementos)."
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "limit",
          "pages"
        ],
        "title": "Page[ReminderRead]"
      },
      "ReminderApplicationSummary": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "position_title": {
            "type": "string",
            "title": "Position Title",
            "examples": [
              "Backend Developer"
            ]
          }
        },
        "type": "object",
        "required": [
          "id",
          "position_title"
        ],
        "title": "ReminderApplicationSummary"
      },
      "ReminderChannel": {
        "type": "string",
        "enum": [
          "in_app"
        ],
        "title": "ReminderChannel",
        "description": "Por dónde se avisa. En el MVP solo existe `IN_APP` (RF-53): es la costura\npara futuros canales (email, Telegram…), que llegarían como una implementación\nnueva de `NotificationChannel` más este valor."
      },
      "ReminderCreate": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title",
            "examples": [
              "Enviar el test técnico"
            ]
          },
          "due_at": {
            "type": "string",
            "format": "date-time",
            "title": "Due At",
            "description": "Fecha límite."
          },
          "application_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Application Id",
            "description": "Solicitud asociada (opcional, RF-50)."
          }
        },
        "type": "object",
        "required": [
          "title",
          "due_at"
        ],
        "title": "ReminderCreate"
      },
      "ReminderRead": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title",
            "examples": [
              "Enviar el test técnico"
            ]
          },
          "due_at": {
            "type": "string",
            "format": "date-time",
            "title": "Due At"
          },
          "application": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReminderApplicationSummary"
              },
              {
                "type": "null"
              }
            ]
          },
          "sent_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sent At",
            "description": "`null` en el MVP: el único canal (in_app) no envía nada (RF-53)."
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "channel": {
            "$ref": "#/components/schemas/ReminderChannel"
          },
          "status": {
            "$ref": "#/components/schemas/ReminderStatus"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "title",
          "due_at",
          "application",
          "sent_at",
          "completed_at",
          "channel",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "ReminderRead"
      },
      "ReminderStatus": {
        "type": "string",
        "enum": [
          "pending",
          "done",
          "dismissed"
        ],
        "title": "ReminderStatus"
      },
      "UnauthorizedError": {
        "properties": {
          "message": {
            "type": "string",
            "const": "unauthorised",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "UnauthorizedError",
        "description": "Respuesta 401 de SuperTokens: no hay sesión, o el token ha caducado o es inválido."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "description": "Access token obtenido en `POST /auth/signin` (cabecera de respuesta `st-access-token`). Caduca a los 5 minutos.",
        "scheme": "bearer"
      },
      "CookieAuth": {
        "type": "apiKey",
        "description": "Cookie httpOnly que usa el navegador tras un login con `st-auth-mode: cookie`. Swagger no puede fijarla a mano.",
        "in": "cookie",
        "name": "sAccessToken"
      }
    }
  },
...
}